On renegotiation, initialize the handshake structure earlier
Initialize the handshake structure when we receive the first handshake
fragment, rather than when we have received the first handshake message.
This is necessary when the first handshake message is fragmented.

This fixes a null pointer dereference when a TLS server renegotiates, but
causes an infinite loop in DTLS clients when they renegotiate.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>