WIP add more detailed defragmentation tests
Assert that we actually have fragments.

This is a work in progress, the tests aren't fully well organized.
Some tests may fail in some configurations because no handhsake message
reaches the minimum size to be fragmented, but we assert that it is. The TLS
1.2 PSK tests fail because they cause ClientHello to be fragmented, but we
don't support receiving a fragmented ClientHello in TLS 1.2.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>