Add PK opaque test cases with an enrollment algorithm
Add test cases where the private key object passed to
`mbedtls_ssl_conf_own_cert()` is a PK opaque key, backed by a PSA key that
has both a primary algorithm and an enrollment algorithm set.

One would naively expect that it doesn't matter whether the algorithm
required by the cipher suite is the primary or the enrollment algorithm. But
it matters to PK for PSA-backed RSA keys, and in turn that makes it matter
for TLS 1.2. The test cases where the required algorithm is the enrollment
algorithm fail.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>